feat: useLogOut 피드백 반영
