+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import { router } from './router/router'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 
-const queryClient = new QueryClient()
+export const queryClient = new QueryClient()
 
 function App() {
     return (
