+++ frontend/src/layouts/RootLayout.tsx
@@ -10,7 +10,6 @@ import { useFetchAndSetUser } from '../hooks/channel/useFetchAndSetUser'
 import { NavbarModalsContainer } from '../pages/auth'
 import { SettingModalContainer } from '../pages/setting/_components/SettingModalContainer'
 import { GenerateErrorModal } from '../pages/report/_components'
-// import AuthWatcher from '../components/AuthWatcher'
 
 export default function RootLayout() {
     const navigate = useNavigate()
@@ -45,7 +44,6 @@ export default function RootLayout() {
 
     return (
         <>
-            {/* <AuthWatcher /> */}
             <NavbarWrapper />
 
             <main
